feat: quest book shows progress
I have made the quests show progress and changed the way the arrows work and some other things; wonder if its alright? Just revert the change if you dont like it; ig.

Though i feel like it would be better if we had a list on the left of all quests and then the steps on the right, or something like that.